Fixes bug when combining discrete and continuous when ensemble missing
A non-missing probability value should always be returned when either
the discrete or continuous model returns a missing value. However this
did not happen when a discrete model that does not use the ensemble
(e.g.  DiscreteConst) is combined with a Continuous model that does use
the ensemble. For sufficiently low CDFs, the valid discrete value is
returned instead of a missing value, because the continuous model was not
checked.